manifest-service/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pymongo==4.3.3
boto3==1.26.90
pydantic==1.10.7
python-dotenv==1.0.0
loguru==0.6.0